+++ app/src/main/java/com/alphawallet/app/ui/widget/entity/ENSHandler.java
@@ -18,7 +18,7 @@ import com.alphawallet.app.C;
 import com.alphawallet.app.R;
 import com.alphawallet.app.repository.TokenRepository;
 import com.alphawallet.app.ui.widget.adapter.AutoCompleteAddressAdapter;
-import com.alphawallet.app.util.AWEnsResolver;
+import com.alphawallet.app.util.ens.AWEnsResolver;
 import com.alphawallet.app.util.Utils;
 import com.alphawallet.app.widget.InputAddress;
 import com.google.gson.Gson;
